+++ data/src/main/kotlin/com/architects/pokearch/data/repository/PokeArchRepository.kt
@@ -21,7 +21,7 @@ class PokeArchRepository @Inject constructor(
         private const val SUBFIX_URL = ".mp3"
     }
 
-    override fun getPokemonTeam() = localDataSource.getPokemonTeam()
+    override fun getPokemonTeam(): Flow<List<PokemonInfo>> = localDataSource.getPokemonTeam()
 
 
     override suspend fun getPokemonList(filter: String, page: Int, limit: Int): List<Pokemon> {
